CRUD rooms

+++ app/Http/Controllers/RoomController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreOrUpdateRoomRequest;
 use App\Models\Room;
 use Exception;
 use Illuminate\Http\Request;
@@ -13,7 +14,8 @@ class RoomController extends Controller
      */
     public function index()
     {
-        //
+        $this->authorize('viewAny', Room::class);
+
         return view('admin/rooms/index', [
             'rooms' => Room::all(),
         ]);
@@ -24,92 +26,65 @@ class RoomController extends Controller
      */
     public function create()
     {
-        //
-        return view('admin/rooms/create');
+        $this->authorize('create', Room::class);
+        $room = new Room([
+            'name' => __('Type the room name here'),
+            'street' => __('Street of the room'),
+        ]);
+        return view('admin/rooms/create', [
+            'room' => $room,
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(StoreOrUpdateRoomRequest $request)
     {
         //
-        $request->validate([
-            'name' => ['string', 'unique:rooms,name',],
-            'street' => ['string'],
-            'city_code' => ['integer', 'max:9999'],
-            'city_name' => ['string'],
-            'building_name' => ['string'],
-            'access_description' => ['string', 'nullable',],
-            'capacity_trainings' => ['integer', 'max:999'],
-            'capacity_matches' => ['integer', 'max:999'],
-        ]);
+        $validated = $request->validated();
 
-        Room::create([
-            'name' => $request->name,
-            'street' => $request->street,
-            'city_code' => $request->city_code,
-            'city_name' => $request->city_name,
-            'building_name' => $request->building_name,
-            'access_description' => $request->access_description,
-            'capacity_trainings' => $request->capacity_trainings,
-            'capacity_matches' => $request->capacity_matches,
-        ]);
+        $room = new Room();
+        
+        $room = Room::create($validated);
 
-        return redirect()->route('rooms.index')->with('success', 'The room ' . $request->name . ' has been added.');
+        return redirect()->route('rooms.index')->with('success', 'The room ' . $room->name . ' has been added.');
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(Room $room)
     {
-        //
+        $this->authorize('view', Room::class);
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(Room $room)
     {
-        //
-        return view('admin.rooms.edit', [
-            'room' => Room::find($id),
+
+        $this->authorize('create', Room::class);
+
+        return view('admin/rooms/edit', [
+            'room' => $room,
         ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(StoreOrUpdateRoomRequest $request, Room $room)
     {
         //
-        $request->validate([
-            'name' => ['string', 'unique:rooms,name,' . $id,],
-            'street' => ['string'],
-            'city_code' => ['integer', 'max:9999'],
-            'city_name' => ['string'],
-            'building_name' => ['string'],
-            'access_description' => ['string', 'nullable',],
-            'capacity_trainings' => ['integer', 'max:999'],
-            'capacity_matches' => ['integer', 'max:999'],
-        ]);
+        $validated = $request->validated();
 
-        $room = Room::find($id);
-
-
-        $room->name = $request->name;
-        $room->street = $request->street;
-        $room->city_code = $request->city_code;
-        $room->city_name = $request->city_name;
-        $room->building_name = $request->building_name;
-        $room->access_description = $request->access_description;
-        $room->capacity_trainings = $request->capacity_trainings;
-        $room->capacity_matches = $request->capacity_matches;
+        $room->fill($validated);
 
         $room->save();
 
-        return redirect()->route('rooms.index')->with('success', 'The room ' . $request->name . ' has been updated.');
+        return redirect()->route('rooms.index')->with('success', 'The room ' . $room->name . ' has been updated.');
     }
 
     /**
@@ -118,12 +93,9 @@ class RoomController extends Controller
      * @param string $id
      * @return void
      */
-    public function destroy(string $id)
+    public function destroy(Room $room)
     {
-        //
-
-        $room = Room::find($id);
-
+        $this->authorize('delete', $room);
         $room->delete();
 
         return redirect()->route('rooms.index')->with('deleted', 'The room ' . $room->name . ' has been deleted.');
@@ -137,18 +109,18 @@ class RoomController extends Controller
      */
     protected function checkCapacity(Request $request): bool
     {
-        $room = Room::find($request->room_id);
-        $requested_capacity = $request->people;
-        $activity = $request->activity;
+        // $room = Room::find($request->room_id);
+        // $requested_capacity = $request->people;
+        // $activity = $request->activity;
 
-        if ($activity == 'training') {
-            $response = $requested_capacity <= $room->capacity_trainings ? true : false;
-            return $response;
-        } elseif ($activity == 'match') {
-            $response = $requested_capacity <= $room->capacity_matches ? true : false;
-            return $response;
-        } else {
-            throw new Exception(__('This activity is unknown. Expected values \'training\' or \'match\'.'));
-        }
+        // if ($activity == 'training') {
+        //     $response = $requested_capacity <= $room->capacity_trainings ? true : false;
+        //     return $response;
+        // } elseif ($activity == 'match') {
+        //     $response = $requested_capacity <= $room->capacity_matches ? true : false;
+        //     return $response;
+        // } else {
+        //     throw new Exception(__('This activity is unknown. Expected values \'training\' or \'match\'.'));
+        // }
     }
 }
